make converters in test-utils like cli.js
