--- fixtures/node-fetch/post_empty.js
@@ -1,5 +1,0 @@
-var fetch = require('node-fetch');
-
-fetch('http://localhost:28139', {
-    method: 'POST'
-});
