+++ fixtures/node-request/get_basic_auth.js
