+++ fixtures/node-request/get_basic_auth_no_user.js
