+++ fixtures/node-request/get_with_all_numerical_data.js
