+++ fixtures/node-request/get_without_headers.js
