+++ fixtures/node-request/put_basic_auth_json_data.js
