+++ fixtures/node/post_empty.js
@@ -1,14 +1,5 @@
-var request = require('request');
+var fetch = require('node-fetch');
 
-var options = {
-    url: 'http://localhost:28139',
+fetch('http://localhost:28139', {
     method: 'POST'
-};
-
-function callback(error, response, body) {
-    if (!error && response.statusCode == 200) {
-        console.log(body);
-    }
-}
-
-request(options, callback);
+});
